+++ src/main/java/com/zepben/evolve/database/sqlite/writers/NetworkCIMWriter.kt
@@ -431,8 +431,8 @@ class NetworkCIMWriter(databaseTables: DatabaseTables) : BaseCIMWriter(databaseT
         val insert = databaseTables.getInsert(TableBatteryUnit::class.java)
 
         insert.setString(table.BATTERY_STATE.queryIndex, batteryUnit.batteryState.name)
-        insert.setDouble(table.RATED_E.queryIndex, batteryUnit.ratedE)
-        insert.setDouble(table.STORED_E.queryIndex, batteryUnit.storedE)
+        insert.setLong(table.RATED_E.queryIndex, batteryUnit.ratedE)
+        insert.setLong(table.STORED_E.queryIndex, batteryUnit.storedE)
 
         return savePowerElectronicsUnit(table, insert, batteryUnit, "battery unit")
     }
